+++ plugins/lang-php/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## [0.2.5] - 2026-07-29
+
+### Fixed
+
+- markdownlint: прибрано порожній рядок наприкінці rules/php/tooling/tooling.mdc
+
+## [0.2.4] - 2026-07-29
+
+### Changed
+
+- Використання diffManifestDeps для порівняння залежностей composer.json
+
 ## [0.2.3] - 2026-07-29
 
 ### Changed